+++ src/Accounts/Accounts/Common/AzureContextModificationCmdlet.cs
@@ -13,11 +13,14 @@
 // ----------------------------------------------------------------------------------
 
 using Microsoft.Azure.Commands.Common.Authentication;
+using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
+using Microsoft.Azure.Commands.Common.Authentication.Core;
 using Microsoft.Azure.Commands.Common.Authentication.Models;
 using Microsoft.Azure.Commands.Common.Authentication.ResourceManager;
 using Microsoft.Azure.Commands.Profile.Properties;
 using Microsoft.Azure.Commands.ResourceManager.Common;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
+using Newtonsoft.Json;
 using System;
 using System.Collections.ObjectModel;
 using System.IO;
@@ -109,7 +112,7 @@ namespace Microsoft.Azure.Commands.Profile.Common
         /// <summary>
         /// Initialize the profile provider based on the autosave setting
         /// </summary>
-        internal void InitializeProfileProvider(bool useAutoSaveProfile = false)
+        internal bool InitializeProfileProvider(bool useAutoSaveProfile = false)
         {
 #if DEBUG
             if (!TestMockSupport.RunningMocked)
@@ -117,9 +120,18 @@ namespace Microsoft.Azure.Commands.Profile.Common
 #endif
                 if (useAutoSaveProfile)
                 {
-                    ProtectedProfileProvider.InitializeResourceManagerProfile();
+                    try
+                    {
+                        ProtectedProfileProvider.InitializeResourceManagerProfile();
+                    }
+                    catch(Exception e)
+                    {
+                        WriteDebugWithTimestamp(e.Message);
+                        ResourceManagerProfileProvider.InitializeResourceManagerProfile(true);
+                    }
                 }
-                else
+
+                if (null == AzureRmProfileProvider.Instance)
                 {
                     switch (GetContextModificationScope())
                     {
@@ -138,6 +150,7 @@ namespace Microsoft.Azure.Commands.Profile.Common
                 ResourceManagerProfileProvider.InitializeResourceManagerProfile();
             }
 #endif
+            return AzureRmProfileProvider.Instance is ProtectedProfileProvider;
         }
 
         /// <summary>
@@ -194,5 +207,36 @@ namespace Microsoft.Azure.Commands.Profile.Common
             return result;
         }
 
+        protected void DisableAutosave(IAzureSession session, bool writeAutoSaveFile, out ContextAutosaveSettings result)
+        {
+            var store = session.DataStore;
+            string tokenPath = Path.Combine(session.TokenCacheDirectory, session.TokenCacheFile);
+            result = new ContextAutosaveSettings
+            {
+                Mode = ContextSaveMode.Process
+            };
+
+            FileUtilities.DataStore = session.DataStore;
+            session.ARMContextSaveMode = ContextSaveMode.Process;
+            var memoryCache = session.TokenCache as AuthenticationStoreTokenCache;
+            if (memoryCache == null)
+            {
+                var diskCache = session.TokenCache as ProtectedFileTokenCache;
+                memoryCache = new AuthenticationStoreTokenCache(new AzureTokenCache());
+                if (diskCache != null && diskCache.Count > 0)
+                {
+                    memoryCache.Deserialize(diskCache.Serialize());
+                }
+
+                session.TokenCache = memoryCache;
+            }
+
+            if (writeAutoSaveFile)
+            {
+                FileUtilities.EnsureDirectoryExists(session.ProfileDirectory);
+                string autoSavePath = Path.Combine(session.ProfileDirectory, ContextAutosaveSettings.AutoSaveSettingsFile);
+                session.DataStore.WriteFile(autoSavePath, JsonConvert.SerializeObject(result));
+            }
+        }
     }
 }